Further code cleanup around String::obtain().

+++ src/main/gram.cpp
@@ -3496,7 +3496,7 @@ static SEXP mkString2(const char *s, size_t len, bool escaped)
     if(known_to_be_latin1) enc = CE_LATIN1;
     else if(!escaped && known_to_be_utf8) enc = CE_UTF8;
 
-    return StringVector::createScalar(String::obtain(s, len, enc));
+    return StringVector::createScalar(String::obtain(s, (int) len, enc));
 }
 
 static SEXP xxdefun(SEXP fname, SEXP formals, SEXP body, YYLTYPE *lloc)
